examples/deployer/requirements.txt
httpx==0.28.1
PyYAML==6.0.2